refactor(tests): consolidate env-copy link tests and collapse overlapping scenarios
