+++ server/tests/integration/sandbox/copy-plan-variants.test.ts
@@ -16,7 +16,7 @@ import { copyProductForOrgs } from "@/internal/products/handlers/handleCopyProdu
 import { ProductService } from "@/internal/products/ProductService.js";
 
 // Copying a single base plan (the "Copy Plan to Production" dialog) must carry
-// the base's variants along, relinked to the newly copied base.
+// the base's variants and license links along, remapped to the target env.
 
 const { db } = initDrizzle();
 const suffix = crypto.randomUUID().slice(0, 8);
@@ -24,19 +24,16 @@ const suffix = crypto.randomUUID().slice(0, 8);
 const BASE_FEATURE = `cpv_base_feat_${suffix}`;
 const VARIANT_FEATURE = `cpv_variant_feat_${suffix}`;
 
-const BASE_PLAN = `cpv_base_${suffix}`;
-const VARIANT_PLAN = `cpv_variant_${suffix}`;
 const RENAME_BASE_PLAN = `cpv_rename_base_${suffix}`;
 const RENAME_TARGET_PLAN = `cpv_rename_target_${suffix}`;
 const RENAME_VARIANT_PLAN = `cpv_rename_variant_${suffix}`;
 const CONFLICT_BASE_PLAN = `cpv_conflict_base_${suffix}`;
 const CONFLICT_TAKEN_PLAN = `cpv_conflict_taken_${suffix}`;
 const CONFLICT_FREE_PLAN = `cpv_conflict_free_${suffix}`;
-const LICENSE_PLAN = `cpv_license_${suffix}`;
 const LICENSE_BASE_PLAN = `cpv_license_base_${suffix}`;
+const REUSED_LICENSE_PLAN = `cpv_reused_license_${suffix}`;
 const PULLED_LICENSE_FEATURE = `cpv_pulled_license_feat_${suffix}`;
 const PULLED_LICENSE_PLAN = `cpv_pulled_license_${suffix}`;
-const PULLED_LICENSE_BASE_PLAN = `cpv_pulled_license_base_${suffix}`;
 
 let org: Organization | undefined;
 
@@ -115,52 +112,16 @@ afterAll(async () => {
 }, 180_000);
 
 describe("copying a single base plan carries its variants", () => {
-	test("copies variants and relinks them to the copied base", async () => {
-		const base = await seedPlan({
-			env: AppEnv.Sandbox,
-			planId: BASE_PLAN,
-			featureIds: [BASE_FEATURE],
-		});
-		await seedPlan({
-			env: AppEnv.Sandbox,
-			planId: VARIANT_PLAN,
-			featureIds: [VARIANT_FEATURE],
-			baseInternalProductId: base.internal_id,
-		});
-
-		await copyPlanToLive({
-			fromProductId: BASE_PLAN,
-			toId: BASE_PLAN,
-			toName: "Copied Base",
-		});
-
-		const livePlans = await listLivePlans();
-		const liveBase = livePlans.find((p) => p.id === BASE_PLAN);
-		const liveVariant = livePlans.find((p) => p.id === VARIANT_PLAN);
-
-		expect(liveBase).toBeDefined();
-		expect(liveVariant).toBeDefined();
-		expect(liveVariant?.base_internal_product_id).toBe(
-			liveBase?.internal_id as string,
-		);
-
-		// The variant's own entitlement feature must come along too.
-		const liveFeatures = await FeatureService.list({
-			db,
-			orgId: org?.id as string,
-			env: AppEnv.Live,
-		});
-		expect(liveFeatures.map((f) => f.id)).toContain(VARIANT_FEATURE);
-	});
-
-	test("renaming the base leaves variant ids and names untouched", async () => {
+	test("copies variants under a renamed base with their features intact", async () => {
 		const base = await seedPlan({
 			env: AppEnv.Sandbox,
 			planId: RENAME_BASE_PLAN,
+			featureIds: [BASE_FEATURE],
 		});
 		const variant = await seedPlan({
 			env: AppEnv.Sandbox,
 			planId: RENAME_VARIANT_PLAN,
+			featureIds: [VARIANT_FEATURE],
 			baseInternalProductId: base.internal_id,
 		});
 
@@ -180,6 +141,14 @@ describe("copying a single base plan carries its variants", () => {
 			liveBase?.internal_id as string,
 		);
 		expect(livePlans.some((p) => p.id === RENAME_BASE_PLAN)).toBe(false);
+
+		// The variant's own entitlement feature must come along too.
+		const liveFeatures = await FeatureService.list({
+			db,
+			orgId: org?.id as string,
+			env: AppEnv.Live,
+		});
+		expect(liveFeatures.map((f) => f.id)).toContain(VARIANT_FEATURE);
 	});
 
 	test("skips a variant whose id is already taken in the target", async () => {
@@ -222,27 +191,35 @@ describe("copying a single base plan carries its variants", () => {
 		expect(liveTaken[0]?.base_internal_product_id).toBeNull();
 	});
 
-	test("license links are recreated against the target's license plan", async () => {
-		const sourceLicense = await seedPlan({
+	test("recreates license links, reusing target plans and pulling absent ones", async () => {
+		await seedFeature({ featureId: PULLED_LICENSE_FEATURE });
+		const reusedSourceLicense = await seedPlan({
 			env: AppEnv.Sandbox,
-			planId: LICENSE_PLAN,
+			planId: REUSED_LICENSE_PLAN,
 		});
-		const targetLicense = await seedPlan({
+		const reusedTargetLicense = await seedPlan({
 			env: AppEnv.Live,
-			planId: LICENSE_PLAN,
+			planId: REUSED_LICENSE_PLAN,
+		});
+		const pulledLicense = await seedPlan({
+			env: AppEnv.Sandbox,
+			planId: PULLED_LICENSE_PLAN,
+			featureIds: [PULLED_LICENSE_FEATURE],
 		});
 		const base = await seedPlan({
 			env: AppEnv.Sandbox,
 			planId: LICENSE_BASE_PLAN,
 		});
-		await planLicenseRepo.upsert({
-			db,
-			parentInternalProductId: base.internal_id,
-			licenseInternalProductId: sourceLicense.internal_id,
-			included: 1,
-			prepaidOnly: true,
-			metadata: {},
-		});
+		for (const license of [reusedSourceLicense, pulledLicense]) {
+			await planLicenseRepo.upsert({
+				db,
+				parentInternalProductId: base.internal_id,
+				licenseInternalProductId: license.internal_id,
+				included: 1,
+				prepaidOnly: true,
+				metadata: {},
+			});
+		}
 
 		await copyPlanToLive({
 			fromProductId: LICENSE_BASE_PLAN,
@@ -252,57 +229,26 @@ describe("copying a single base plan carries its variants", () => {
 
 		const livePlans = await listLivePlans();
 		const liveBase = livePlans.find((p) => p.id === LICENSE_BASE_PLAN);
+		const livePulledLicense = livePlans.find(
+			(p) => p.id === PULLED_LICENSE_PLAN,
+		);
 		const liveLinks = await planLicenseRepo.listWithLicensePlanIdByParents({
 			db,
 			parentInternalProductIds: [liveBase?.internal_id as string],
 		});
 
-		expect(liveLinks.length).toBe(1);
-		expect(liveLinks[0]?.licensePlanId).toBe(LICENSE_PLAN);
-		expect(liveLinks[0]?.planLicense.license_internal_product_id).toBe(
-			targetLicense.internal_id,
-		);
-	});
-
-	test("a license plan absent from the target is copied and linked", async () => {
-		await seedFeature({ featureId: PULLED_LICENSE_FEATURE });
-		const sourceLicense = await seedPlan({
-			env: AppEnv.Sandbox,
-			planId: PULLED_LICENSE_PLAN,
-			featureIds: [PULLED_LICENSE_FEATURE],
-		});
-		const base = await seedPlan({
-			env: AppEnv.Sandbox,
-			planId: PULLED_LICENSE_BASE_PLAN,
-		});
-		await planLicenseRepo.upsert({
-			db,
-			parentInternalProductId: base.internal_id,
-			licenseInternalProductId: sourceLicense.internal_id,
-			included: 2,
-			prepaidOnly: true,
-			metadata: {},
-		});
-
-		await copyPlanToLive({
-			fromProductId: PULLED_LICENSE_BASE_PLAN,
-			toId: PULLED_LICENSE_BASE_PLAN,
-			toName: "Pulled License Base",
-		});
-
-		const livePlans = await listLivePlans();
-		const liveBase = livePlans.find((p) => p.id === PULLED_LICENSE_BASE_PLAN);
-		const liveLicense = livePlans.find((p) => p.id === PULLED_LICENSE_PLAN);
-		const liveLinks = await planLicenseRepo.listWithLicensePlanIdByParents({
-			db,
-			parentInternalProductIds: [liveBase?.internal_id as string],
-		});
-
-		expect(liveLicense).toBeDefined();
-		expect(liveLinks.length).toBe(1);
-		expect(liveLinks[0]?.planLicense.license_internal_product_id).toBe(
-			liveLicense?.internal_id as string,
-		);
+		// The reused link resolves to the pre-existing target license; the pulled
+		// license is copied in and linked.
+		expect(liveLinks).toHaveLength(2);
+		expect(
+			liveLinks.find((link) => link.licensePlanId === REUSED_LICENSE_PLAN)
+				?.planLicense.license_internal_product_id,
+		).toBe(reusedTargetLicense.internal_id);
+		expect(livePulledLicense).toBeDefined();
+		expect(
+			liveLinks.find((link) => link.licensePlanId === PULLED_LICENSE_PLAN)
+				?.planLicense.license_internal_product_id,
+		).toBe(livePulledLicense?.internal_id as string);
 
 		const liveFeatures = await FeatureService.list({
 			db,
